+++ intelmq/bots/parsers/spamhaus/parser_cert.py
@@ -136,6 +136,16 @@ class SpamhausCERTParserBot(ParserBot):
                 if row_splitted[7] == '25':
                     event.add('protocol.application', 'smtp')
                 event.add('event_description.text', 'This device sent extortion mails.')
+            elif malware == 'misc':
+                # According to info from abuseat.org
+                # exact malware is not known, but it is known to be a botnet
+                event.add('classification.type', 'spam')
+                event.add('classification.identifier', 'spam')
+                if row_splitted[7] == '25':
+                    event.add('protocol.application', 'smtp')
+                event.add('event_description.text', 'This device is sending spam as part of a botnet. See abuseat.org for more details.')
+                event.add('event_description.url', 'https://www.abuseat.org/lookup.cgi')
+                event.add('feed.documentation', 'https://www.abuseat.org/lookup.cgi', overwrite=False)
             else:
                 if malware == 'auto':
                     malware = 's_other'